Consolidate the two overlapping Keyring fixes into one coherent change so only
this PR needs to land:

- #743 (Seranged): scoped credential validation (isCredentialForContext),
  unified extension-state handling (handleExtensionState — wrong wallet/policy/
  chain, error/proving states, incomplete manual checks), isCheckingStatus /
  statusMessage UX, and keeping the gate in Loading while hook metadata resolves.
- This branch: the native->integrator getter-name fallback
  (policyId/keyring -> getPolicyId/getKeyring) that fixes the actual
  "Start Verification does nothing" root cause, plus an explicit error state when
  the policy cannot be resolved.

Conflict resolutions:
- useKeyring: keep both isCredentialForContext and readHookTargetValue; drop the
  now-unused readHookTargetField; checkCredential reads via the getter fallback;
  launchExtension keeps the missing-prerequisites logging.
- getPolicyId modeled as uint32 (matching the verified integrator ABI and the
  SDK-side fix), so no Number() coercion is needed.
- Combined the useKeyring test suite: #743's state-machine tests plus the
  getter-name fallback tests.

+++ abis/keyring.ts
@@ -24,11 +24,12 @@ export const keyringHookTargetAbi = [
   // Integrator-supplied hook targets (e.g. HookTargetAccessControlKeyringUnwind)
   // expose the same values behind `get`-prefixed getters instead of the public
   // immutables above. useKeyring falls back to these when policyId()/keyring()
-  // revert. Declared uint256 as a safe superset — the value is coerced to Number.
+  // revert. getPolicyId() is uint32 on the verified integrator ABI, matching the
+  // native policyId() getter (so viem returns a plain number, no coercion).
   {
     inputs: [],
     name: 'getPolicyId',
-    outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
+    outputs: [{ internalType: 'uint32', name: '', type: 'uint32' }],
     stateMutability: 'view',
     type: 'function',
   },
